Refactoring: Rename class names
