+++ example.js
@@ -1,4 +1,4 @@
-import {TableConstructor} from "./src/TableConstructor";
+import {TableConstructor} from "./src/tableConstructor";
 
 let table = document.getElementsByClassName("main_wrapper")[0];
 let table_editor = new TableConstructor({width: "800px"});
