+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="style.css">
-    <link rel="stylesheet" href="src/TableConstructor.css">
+    <link rel="stylesheet" href="src/tableConstructor.css">
     <title>Table constructor</title>
 </head>
 <body>
