+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "TableConstructor",
   "version": "0.0.2",
-  "main": "src/TableConstructor.js",
+  "main": "src/tableConstructor.js",
   "repository": "https://github.com/horoyami/TableConstructor.git",
   "author": "Vinogradov Victor <re.horoyami.not@gmail.com>",
   "devDependencies": {
