--- src/Border.css
@@ -1,12 +1,0 @@
-.b-add_button {
-    height: 19px;
-    width: 19px;
-}
-
-.b-bor {
-    background: #368be6;
-}
-
-.b-hidden {
-    visibility: hidden;
-}
