+++ src/borderMenu.css
@@ -1,0 +1,14 @@
+/* TCM - project's prefix*/
+
+.TCM__border-menu__add-button {
+    height: 19px;
+    width: 19px;
+}
+
+.TCM__border-menu {
+    background: #368be6;
+}
+
+.TCM__border-menu_hidden {
+    visibility: hidden;
+}
