+++ src/horizontalBorderMenu.css
@@ -1,4 +1,6 @@
-.hb-hor_wrapper {
+/* TCM - project's prefix*/
+
+.TCM__border-menu_horizontal {
     width: 100%;
     height: 21px;
     display: flex;
@@ -9,7 +11,7 @@
     top: 0px;
 }
 
-.hb-hor {
+.TCM__border-menu__1pxline_horizontal {
     min-height: 1px;
     width: 100%;
     margin-right: 10px;
