+++ style.css
@@ -13,7 +13,7 @@ html, body {
     align-items: center;
 }
 
-.editable_table {
+.TCM__editable-table {
     width: 80px;
     height: 80px;
 }
